Python Wrapper: Implement Object
